Add a basic searcher
The searcher only scores based on number of found trigrams. However, there's a
foundation in place to make this better - both by using the current position,
word index, and the previous position & word index. This is untested.

More importantly, added a struct to track top scores and returned sorted entries.